+++ src/net_main.cpp
@@ -2,6 +2,7 @@
 
 #include "quakedef.h"
 #include "net_vcr.h"
+#include <stdint.h>
 
 qsocket_t* net_activeSockets = NULL;
 qsocket_t* net_freeSockets = NULL;
@@ -454,7 +455,7 @@ NET_CheckNewConnections
 struct {
     double time;
     int op;
-    long session;
+    intptr_t session; // Changed from long to intptr_t
 } vcrConnect;
 
 qsocket_t* NET_CheckNewConnections(void)
@@ -478,7 +479,7 @@ qsocket_t* NET_CheckNewConnections(void)
             if (recording) {
                 vcrConnect.time = host_time;
                 vcrConnect.op = VCR_OP_CONNECT;
-                vcrConnect.session = (long)ret;
+                vcrConnect.session = (intptr_t)ret;
                 Sys_FileWrite(vcrFile, &vcrConnect, sizeof(vcrConnect));
                 Sys_FileWrite(vcrFile, ret->address, NET_NAMELEN);
             }
@@ -535,7 +536,7 @@ returns -1 if connection is invalid
 struct {
     double time;
     int op;
-    long session;
+    intptr_t session; // Changed from long to intptr_t
     int ret;
     int len;
 } vcrGetMessage;
@@ -582,7 +583,7 @@ int NET_GetMessage(qsocket_t* sock)
         if (recording) {
             vcrGetMessage.time = host_time;
             vcrGetMessage.op = VCR_OP_GETMESSAGE;
-            vcrGetMessage.session = (long)sock;
+            vcrGetMessage.session = (intptr_t)sock;
             vcrGetMessage.ret = ret;
             vcrGetMessage.len = net_message.cursize;
             Sys_FileWrite(vcrFile, &vcrGetMessage, 24);
@@ -592,7 +593,7 @@ int NET_GetMessage(qsocket_t* sock)
         if (recording) {
             vcrGetMessage.time = host_time;
             vcrGetMessage.op = VCR_OP_GETMESSAGE;
-            vcrGetMessage.session = (long)sock;
+            vcrGetMessage.session = (intptr_t)sock;
             vcrGetMessage.ret = ret;
             Sys_FileWrite(vcrFile, &vcrGetMessage, 20);
         }
@@ -615,7 +616,7 @@ returns -1 if the connection died
 struct {
     double time;
     int op;
-    long session;
+    intptr_t session; // Changed from long to intptr_t
     int r;
 } vcrSendMessage;
 
@@ -642,7 +643,7 @@ int NET_SendMessage(qsocket_t* sock, sizebuf_t* data)
     if (recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (intptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
@@ -673,7 +674,7 @@ int NET_SendUnreliableMessage(qsocket_t* sock, sizebuf_t* data)
     if (recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_SENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (intptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
@@ -708,7 +709,7 @@ qboolean NET_CanSendMessage(qsocket_t* sock)
     if (recording) {
         vcrSendMessage.time = host_time;
         vcrSendMessage.op = VCR_OP_CANSENDMESSAGE;
-        vcrSendMessage.session = (long)sock;
+        vcrSendMessage.session = (intptr_t)sock;
         vcrSendMessage.r = r;
         Sys_FileWrite(vcrFile, &vcrSendMessage, 20);
     }
